backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
bcrypt
python-dotenv
Flask-Migrate
psycopg2-binary
prometheus-flask-exporter
